requirements.txt
streamlit
opencv-python
numpy
pillow
scikit-learn 
pandas 
joblib
streamlit-drawable-canvas